Extract format_duration into a shared reporter helper
The summary, llm, and text reporters all carried their own copy of
the duration formatter, with the recent minutes:seconds change landing
in only two of them. Move the function into a new tryke_reporter::duration
module so all three reporters share one implementation and per-test
durations in the text reporter pick up the M:SS.SS format too.
