+++ crates/tryke_reporter/src/summary.rs
@@ -1,9 +1,9 @@
 use std::io;
-use std::time::Duration;
 
 use owo_colors::OwoColorize;
 use tryke_types::{RunSummary, TestItem};
 
+use crate::duration::format_duration;
 use crate::reporter::WatchIdleInfo;
 
 /// Keyboard shortcuts shown beneath the summary/idle badge in watch
@@ -21,26 +21,6 @@ fn write_watch_keybindings<W: io::Write>(writer: &mut W) {
     for (key, label) in WATCH_KEYBINDINGS {
         let pad = " ".repeat(LABEL_COLUMN_WIDTH.saturating_sub(key.len()));
         let _ = writeln!(writer, "{pad}{}  {}", (*key).bold(), (*label).dimmed());
-    }
-}
-
-fn format_duration(d: Duration) -> String {
-    let secs = d.as_secs_f64();
-    if secs < 1.0 {
-        format!("{:.2}ms", secs * 1000.0)
-    } else if secs < 60.0 {
-        format!("{secs:.2}s")
-    } else {
-        // Integer math avoids the float-cast precision lint. The `+ 5`
-        // rounds to the nearest centisecond so output matches the
-        // sub-minute branches (which round via `:.2`); collapsing
-        // everything into centiseconds first lets carry through
-        // seconds/minutes happen naturally on decomposition.
-        let centis = (d.as_millis() + 5) / 10;
-        let minutes = centis / 6000;
-        let secs = (centis / 100) % 60;
-        let cs = centis % 100;
-        format!("{minutes}:{secs:02}.{cs:02}")
     }
 }
